Linear Search
Linear search implementation in C++ using Iterative and recursive approach.
Linear Search: It is the simplest search algorithm with an time complexity of O(n) where n is the size of the array. This is a brute force approach where we traverse through the array and compare each element with the required search element. If they are equal, we return the index at which the match was found. If no match is found, we return -1 and indicate that the element wasn't present in the array.